+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## v2.3.2 - 2019-03-26
 ### Fixed
 * Prevent users from being able to enter room assignments twice ([#950](https://gitlab.com/yale-sdmp/vesta/issues/950)).
 * Fix DrawlessGroupsController not correctly redirecting if the group has a draw ([#920](https://gitlab.com/yale-sdmp/vesta/issues/920)).